correct error item

+++ src/fr/edminecoreteam/corepractice/listeners/JoinEvent.java
@@ -24,9 +24,9 @@ public class JoinEvent implements Listener
         }
 
         Location lobbySpawn = new Location(Bukkit.getWorld(core.getConfig().getString("Lobby.world")),
-                (float) core.getConfig().getDouble("Lobby.world.x")
-                , (float) core.getConfig().getDouble("Lobby.world.y")
-                , (float) core.getConfig().getDouble("Lobby.world.z")
+                core.getConfig().getDouble("Lobby.world.x")
+                , core.getConfig().getDouble("Lobby.world.y")
+                , core.getConfig().getDouble("Lobby.world.z")
                 , (float) core.getConfig().getDouble("Lobby.world.t")
                 , (float) core.getConfig().getDouble("Lobby.world.b"));
